+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.16.2] - 2024-08-23
+
 ### Changed
 
 - The error message in some edge cases involving complex generics and abstract classes is now improved. ([Issue 983](https://github.com/jqno/equalsverifier/issues/983))
@@ -1228,7 +1230,8 @@ Please don't use version 1.3; [it's a broken release](https://jqno.nl/post/2013/
 
 You can now use EqualsVerifier!
 
-[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.16.1...HEAD
+[unreleased]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.16.2...HEAD
+[3.16.2]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.16.1...equalsverifier-3.16.2
 [3.16.1]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.16...equalsverifier-3.16.1
 [3.16]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.8...equalsverifier-3.16
 [3.15.8]: https://github.com/jqno/equalsverifier/compare/equalsverifier-3.15.7...equalsverifier-3.15.8